buy ticketes, login fix, content update, session storage instead of local
