+++ public/content/footer/content.json
@@ -43,7 +43,7 @@
           "hyperlink": "/"
         },
         {
-          "title": "Terms &amp; Conditions",
+          "title": "Terms & Conditions",
           "hyperlink": "/"
         },
         {
